refactor: client package
