+++ pkg/client/simple/auth.go
@@ -1,4 +1,4 @@
-package client
+package simple
 
 import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth" // package needed for auth providers like GCP
