+++ pkg/runner/kubectl/wait_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
-	"github.com/kyverno/chainsaw/pkg/client"
+	"github.com/kyverno/chainsaw/pkg/client/simple"
 	restutils "github.com/kyverno/chainsaw/pkg/utils/rest"
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -16,7 +16,7 @@ import (
 func TestWait(t *testing.T) {
 	config, err := restutils.DefaultConfig(clientcmd.ConfigOverrides{})
 	assert.NoError(t, err)
-	client, err := client.New(config)
+	client, err := simple.New(config)
 	assert.NoError(t, err)
 	tests := []struct {
 		name    string
